RestrictedBoltzmannMachines.jl Documentation

Restricted Boltzmann Machines

We consider a restricted Boltzmann machine (RBM) with visible units $\mathbf{v} = (v_1, \ldots, v_N)$ and hidden units $\mathbf{h} = (h_1, \ldots, h_M)$. The energy function is given by:

\[E(\mathbf{v}, \mathbf{h}) = \sum_i \mathcal{V}_i(v_i) + \sum_{\mu}\mathcal{U}_{\mu}(h_{\mu}) - \sum_{i\mu} w_{i\mu} v_i h_{\mu}\]

where $\mathcal{V}_i(v_i)$ and $\mathcal{U}_{\mu}(h_{\mu})$ are the unit potentials and $w_{i\mu}$ the interaction weights. The probability of a configuration is

\[P(\mathbf{v}, \mathbf{h}) = \frac{1}{Z}\mathrm{e}^{-\beta E(\mathbf{v},\mathbf{h})}\]

where

\[Z = \sum_{\mathbf{v}, \mathbf{h}} \mathrm{e}^{-\beta E(\mathbf{v}, \mathbf{h})}\]

is the partition function and $\beta$ the inverse temperature. The machine assigns a likelihood:

\[P(\mathbf{v}) = \underset{\mathbf{h}}{\sum} P (\mathbf{v}, \mathbf{h}) = \frac{1}{Z} \mathrm{e}^{-\beta E_{\textrm{eff}}(\mathbf{v})}\]

to visible configurations, where $E_{\textrm{eff}}(\mathbf{v})$ is the free energy:

\[E_{\textrm{eff}}(\mathbf{v}) = \sum_i \mathcal{V}_i(v_i) - \sum_{\mu} \Gamma_{\mu} \left(\sum_i w_{i \mu} v_i \right)\]

and

\[\Gamma_{\mu}(I) = \frac{1}{\beta} \ln \sum_{h_\mu} \mathrm{e}^{\beta(I h_{\mu} - \mathcal{U}_{\mu}(h_{\mu}))}\]

are the cumulant generating functions associated to the hidden unit potentials.

Note that $\beta$ refers to the inverse temperature in the distribution $P(\mathbf{v},\mathbf{h})$. If instead we want to sample the marginal $P(\mathbf{v})$ at a different inverse temperature $\beta_v$, we would have to use the distribution:

\[P_{\beta_v}(\mathbf{v}) = \frac{\mathrm{e}^{- \beta_v E_\textrm{eff} (\mathbf{v})}}{\sum_{\mathbf{v}} \mathrm{e}^{-\beta_{\mathrm{v}} E_{\textrm{eff}} (\mathbf{v})}}\]

Reference

This package doesn't export any symbols. To make names shorter, we will import the package as RBMs:

import RestrictedBoltzmannMachines as RBMs

RBMs

RestrictedBoltzmannMachines.RBM — Type

RBM

Represents a restricted Boltzmann Machine.

RestrictedBoltzmannMachines.energy — Function

energy(rbm, v, h)

Energy of the rbm in the configuration (v,h).

RestrictedBoltzmannMachines.free_energy — Function

free_energy(rbm, v, β=1)

Free energy of visible configuration (after marginalizing hidden configurations).

RestrictedBoltzmannMachines.interaction_energy — Function

interaction_energy(rbm, v, h)

Weight mediated interaction energy.

RestrictedBoltzmannMachines.inputs_v_to_h — Function

inputs_v_to_h(rbm, v)

Interaction inputs from visible to hidden layer.

RestrictedBoltzmannMachines.inputs_h_to_v — Function

inputs_h_to_v(rbm, h)

Interaction inputs from hidden to visible layer.

RestrictedBoltzmannMachines.sample_h_from_v — Function

sample_h_from_v(rbm, v, β=1)

Samples a hidden configuration conditional on the visible configuration v.

RestrictedBoltzmannMachines.sample_v_from_h — Function

sample_v_from_h(rbm, h, β = 1)

Samples a visible configuration conditional on the hidden configuration h.

RestrictedBoltzmannMachines.sample_v_from_v — Function

sample_v_from_v(rbm, v, β = 1; steps = 1)

Samples a visible configuration conditional on another visible configuration v.

RestrictedBoltzmannMachines.sample_h_from_h — Function

sample_h_from_h(rbm, h, β = 1; steps = 1)

Samples a hidden configuration conditional on another hidden configuration h.

Layers

RestrictedBoltzmannMachines.Gaussian — Type

Gaussian(θ, γ)

Gaussian layer, with location parameters θ and scale parameters γ.