RestrictedBoltzmannMachines.jl Reference

RestrictedBoltzmannMachines.Binary — Type

Binary(θ)

Binary layer, with external fields θ.

RestrictedBoltzmannMachines.Gaussian — Type

Gaussian(θ, γ)

Gaussian layer, with location parameters θ and scale parameters γ.

RestrictedBoltzmannMachines.Potts — Type

Potts(θ)

Potts layer, with external fields θ. Encodes categorical variables as one-hot vectors. The number of classes is the size of the first dimension.

RestrictedBoltzmannMachines.RBM — Type

RBM

Represents a restricted Boltzmann Machine.

RestrictedBoltzmannMachines.ReLU — Type

ReLU(θ, γ)

ReLU layer, with location parameters θ and scale parameters γ.

RestrictedBoltzmannMachines.Spin — Type

Spin(θ)

Spin layer, with external fields θ.

RestrictedBoltzmannMachines.batchmul — Method

batchmul(fields, x)

Forms the product fields * x, summing over the fields dimension while leaving the batch dimensions.

RestrictedBoltzmannMachines.categorical_rand — Method

categorical_rand(ps)

Randomly draw i with probability ps[i]. You must ensure that ps defines a proper probability distribution.

RestrictedBoltzmannMachines.categorical_sample — Method

categorical_sample(P)

Given a probability array P of size (q, *), returns an array C of size (*), such that C[i] ∈ 1:q is a random sample from the categorical distribution P[:,i]. You must ensure that P defines a proper probability distribution.

RestrictedBoltzmannMachines.categorical_sample_from_logits — Method

categorical_sample_from_logits(logits)

Given a logits array logits of size (q, *) (where q is the number of classes), returns an array X of size (*), such that X[i] is a categorical random sample from the distribution with logits logits[:,i].

RestrictedBoltzmannMachines.categorical_sample_from_logits_gumbel — Method

categorical_sample_from_logits_gumbel(logits)

Like categoricalsamplefrom_logits, but using the Gumbel trick.

RestrictedBoltzmannMachines.energy — Method

energy(rbm, v, h)

Energy of the rbm in the configuration (v,h).

RestrictedBoltzmannMachines.flip_layers — Method

flip_layers(rbm)

Returns a new RBM with viible and hidden layers flipped.

RestrictedBoltzmannMachines.free_energy — Function

free_energy(rbm, v, β=1)

Free energy of visible configuration (after marginalizing hidden configurations).

RestrictedBoltzmannMachines.generate_sequences — Function

generate_sequences(n, A = 0:1)

Retruns an iterator over all sequences of length n out of the alphabet A.

RestrictedBoltzmannMachines.init! — Method

init!(rbm, data; ϵ = 1e-6)

Inits the RBM, computing average visible unit activities from data.

RestrictedBoltzmannMachines.init_weights! — Method

init_weights!(rbm; w=1)

Random initialization of weights, as independent normals with variance 1/N. All patterns are of norm 1.

RestrictedBoltzmannMachines.inputs_h_to_v — Method

inputs_h_to_v(rbm, h)

Interaction inputs from hidden to visible layer.

RestrictedBoltzmannMachines.inputs_v_to_h — Method

inputs_v_to_h(rbm, v)

Interaction inputs from visible to hidden layer.

RestrictedBoltzmannMachines.interaction_energy — Method

interaction_energy(rbm, v, h)

Weight mediated interaction energy.

RestrictedBoltzmannMachines.log_likelihood — Function

log_likelihood(rbm, v, β=1)

Log-likelihood of v under rbm, with the partition function compued by extensive enumeration. For discrete layers, this is exponentially slow for large machines.

RestrictedBoltzmannMachines.log_partition — Function

log_partition(rbm, β = 1)

Log-partition of the rbm at inverse temperature β, computed by extensive enumeration of states (except for particular cases such as Gaussian-Gaussian) RBM). This is exponentially slow for large machines.

RestrictedBoltzmannMachines.log_pseudolikelihood — Function

log_pseudolikelihood(rbm, v, sites, β=1)

Log-pseudolikelihood of a site conditioned on the other sites, where sites is an array of site indices (CartesianIndex), one for each batch. Returns an array of log-pseudolikelihood, for each batch.

RestrictedBoltzmannMachines.log_pseudolikelihood — Function

log_pseudolikelihood_rand(rbm, v, β=1)

Log-pseudolikelihood of randomly chosen sites conditioned on the other sites. For each configuration choses a samplefrominputs site, and returns the mean of the computed pseudo-likelihoods.

RestrictedBoltzmannMachines.log_site_trace — Function

log_site_trace(site, rbm, v, β=1)

Log of the trace over configurations of site. Here v must consist of a single batch.

RestrictedBoltzmannMachines.log_site_traces — Function

log_site_traces(rbm, v, sites, β=1)

Log of the trace over configurations of sites, where sites is an array of site indices (CartesianIndex), for each batch. Returns an array of the log-traces for each batch.

RestrictedBoltzmannMachines.lognormcdf — Method

lognormcdf(a, b)

Computes log(normcdf(a, b)), but retaining accuracy.

RestrictedBoltzmannMachines.mean_ — Method

mean_(A, dims)

Takes the mean of A across dimensions dims and drops them.

RestrictedBoltzmannMachines.mills — Method

mills(x::Real)

Mills ratio of the standard normal distribution. Defined as (1 - cdf(x)) / pdf(x).

RestrictedBoltzmannMachines.minibatch_count — Method

minibatch_count(nobs; batchsize)

Number of minibatches.

RestrictedBoltzmannMachines.minibatch_count — Method

minibatch_count(data; batchsize)

Number of minibatches.

RestrictedBoltzmannMachines.minibatches — Method

minibatches(nobs; batchsize, shuffle = true)

Partition nobs into minibatches of length n. If necessary repeats some observations to complete last batches. (Therefore all batches are of the same size n).

RestrictedBoltzmannMachines.minibatches — Method

minibatches(datas...; batchsize)

Splits the given datas into minibatches. Each minibatch is a tuple where each entry is a minibatch from the corresponding data within datas. All minibatches are of the same size batchsize (if necessary repeating some samples at the last minibatches).

RestrictedBoltzmannMachines.normalize_weights — Method

normalize_weights(rbm)

Rescales weights so that norm(w[:,μ]) = 1 for all μ (making individual weights ~ 1/√N). The resulting RBM shares layer parameters with the original, but weights are a new array.

RestrictedBoltzmannMachines.normcdf — Method

normcdf(a, b)

Probablity that a ≤ Z ≤ b, where Z is a standard normal samplefrominputs variable. WARNING: Silently returns a negative value if a > b.

RestrictedBoltzmannMachines.normcdf — Method

normcdf(x)

Probablity that Z ≤ x, where Z is a standard normal samplefrominputs variable.

RestrictedBoltzmannMachines.normcdfinv — Method

normcdfinv(x)

Inverse of normcdf.

RestrictedBoltzmannMachines.onehot_decode — Method

onehot_decode(X)

Given a onehot encoded array X of N + 1 dimensions, returns the equivalent categorical array of N dimensions.

RestrictedBoltzmannMachines.onehot_encode — Method

onehot_encode(A, code)

Given an array A of N dimensions, returns a one-hot encoded array of N + 1 dimensions where single entries of the first dimension are one.

RestrictedBoltzmannMachines.randgumbel — Method

randgumbel(T = Float64)

Generates a random Gumbel variate.

RestrictedBoltzmannMachines.randnt — Method

randnt([rng], a)

Random standard normal lower truncated at a (that is, Z ≥ a).

RestrictedBoltzmannMachines.randnt_half — Method

randnt_half([rng], μ, σ)

Samples the normal distribution with mean μ and standard deviation σ truncated to positive values.

RestrictedBoltzmannMachines.reconstruction_error — Function

reconstruction_error(rbm, v, β = 1; steps = 1)

Stochastic reconstruction error of v.

RestrictedBoltzmannMachines.sample_h_from_h — Function

sample_h_from_h(rbm, h, β = 1; steps = 1)

Samples a hidden configuration conditional on another hidden configuration h.

RestrictedBoltzmannMachines.sample_h_from_v — Function

sample_h_from_v(rbm, v, β=1)

Samples a hidden configuration conditional on the visible configuration v.

RestrictedBoltzmannMachines.sample_v_from_h — Function

sample_v_from_h(rbm, h, β = 1)

Samples a visible configuration conditional on the hidden configuration h.

RestrictedBoltzmannMachines.sample_v_from_v — Function

sample_v_from_v(rbm, v, β = 1; steps = 1)

Samples a visible configuration conditional on another visible configuration v.

RestrictedBoltzmannMachines.sqrt1half — Method

sqrt1half(x)

Accurate computation of sqrt(1 + (x/2)^2) + |x|/2.

RestrictedBoltzmannMachines.sum_ — Method

sum_(A, dims)

Sums A over dimensions dims and drops them.

RestrictedBoltzmannMachines.tnmean — Method

tnmean(a)

Mean of the standard normal distribution, truncated to the interval (a, +∞).

RestrictedBoltzmannMachines.tnstd — Method

tnstd(a)

Standard deviation of the standard normal distribution, truncated to the interval (a, +∞).

RestrictedBoltzmannMachines.tnvar — Method

tnvar(a)

Variance of the standard normal distribution, truncated to the interval (a, +∞). WARNING: Fails for very very large values of a.

RestrictedBoltzmannMachines.train! — Method

train!(rbm, data)

Trains the RBM on data.

RestrictedBoltzmannMachines.tuplen — Method

tuplen(Val(N))

Constructs the tuple (1, 2, ..., N).

RestrictedBoltzmannMachines.unwhiten — Method

unwhiten(rbm, data)

Given an RBM trained on whitened data, returns an RBM that can look at original data.

RestrictedBoltzmannMachines.weighted_mean — Method

weighted_mean(v, w)

Mean of v with weights w.