RestrictedBoltzmannMachines.jl Reference

RestrictedBoltzmannMachines.Binary — Type

Binary(θ)

Binary layer, with external fields θ.

RestrictedBoltzmannMachines.Gaussian — Type

Gaussian(θ, γ)

Gaussian layer, with location parameters θ and scale parameters γ.

RestrictedBoltzmannMachines.Potts — Type

Potts(θ)

Potts layer, with external fields θ. Encodes categorical variables as one-hot vectors. The number of classes is the size of the first dimension.

RestrictedBoltzmannMachines.RBM — Type

RBM

Represents a restricted Boltzmann Machine.

RestrictedBoltzmannMachines.ReLU — Type

ReLU(θ, γ)

ReLU layer, with location parameters θ and scale parameters γ.

RestrictedBoltzmannMachines.Spin — Type

Spin(θ)

Spin layer, with external fields θ. The energy of a layer with units $s_i$ is given by:

\[E = -\sum_i \theta_i s_i\]

where each spin $s_i$ takes values $\pm 1$.

RestrictedBoltzmannMachines.block_matrix_invert — Method

block_matrix_invert(A, B, C, D)

Inversion of a block matrix, using the formula:

\[\begin{bmatrix} \mathbf{A} & \mathbf{B} \\ \mathbf{C} & \mathbf{D} \end{bmatrix}^{-1} = \begin{bmatrix} \left(\mathbf{A} - \mathbf{B} \mathbf{D}^{-1} \mathbf{C}\right)^{-1} & \mathbf{0} \\ \mathbf{0} & \left(\mathbf{D} - \mathbf{C} \mathbf{A}^{-1} \mathbf{B}\right)^{-1} \end{bmatrix} \begin{bmatrix} \mathbf{I} & -\mathbf{B} \mathbf{D}^{-1} \\ -\mathbf{C} \mathbf{A}^{-1} & \mathbf{I} \end{bmatrix}\]

Note that this assumes that A and D are both invertible.

RestrictedBoltzmannMachines.block_matrix_logdet — Method

block_matrix_logdet(A, B, C, D)

Log-determinant of a block matrix using the determinant lemma.

\[\det\left( \begin{bmatrix} \mathbf{A} & \mathbf{B} \\ \mathbf{C} & \mathbf{D} \end{bmatrix} \right) = \det(A) \det(D - CA^{-1}B) = \det(D) \det(A - BD^{-1}C)\]

Here we assume that A and D are invertible, and moreover are easy to invert (for example, if they are diagonal). We use this to chose one or the other of the two formulas above.

RestrictedBoltzmannMachines.categorical_rand — Method

categorical_rand(ps)

Randomly draw i with probability ps[i]. You must ensure that ps defines a proper probability distribution.

RestrictedBoltzmannMachines.categorical_sample — Method

categorical_sample(P)

Given a probability array P of size (q, *), returns an array C of size (*), such that C[i] ∈ 1:q is a random sample from the categorical distribution P[:,i]. You must ensure that P defines a proper probability distribution.

RestrictedBoltzmannMachines.categorical_sample_from_logits — Method

categorical_sample_from_logits(logits)

Given a logits array logits of size (q, *) (where q is the number of classes), returns an array X of size (*), such that X[i] is a categorical random sample from the distribution with logits logits[:,i].

RestrictedBoltzmannMachines.categorical_sample_from_logits_gumbel — Method

categorical_sample_from_logits_gumbel(logits)

Like categoricalsamplefrom_logits, but using the Gumbel trick.

RestrictedBoltzmannMachines.energy — Method

energy(rbm, v, h)

Energy of the rbm in the configuration (v,h).

RestrictedBoltzmannMachines.flip_layers — Method

flip_layers(rbm)

Returns a new RBM with viible and hidden layers flipped.

RestrictedBoltzmannMachines.free_energy — Function

free_energy(rbm, v, β=1)

Free energy of visible configuration (after marginalizing hidden configurations).

RestrictedBoltzmannMachines.generate_sequences — Function

generate_sequences(n, A = 0:1)

Retruns an iterator over all sequences of length n out of the alphabet A.

RestrictedBoltzmannMachines.initialize! — Function

initialize!(rbm, [data]; ϵ = 1e-6)

Initializes the RBM parameters. If provided, matches average visible unit activities from data.

initialize!(layer, [data]; ϵ = 1e-6)

Initializes a layer. If provided, matches average unit activities from data.

RestrictedBoltzmannMachines.inputs_h_to_v — Method

inputs_h_to_v(rbm, h)

Interaction inputs from hidden to visible layer.

RestrictedBoltzmannMachines.inputs_v_to_h — Method

inputs_v_to_h(rbm, v)

Interaction inputs from visible to hidden layer.

RestrictedBoltzmannMachines.interaction_energy — Method

interaction_energy(rbm, v, h)

Weight mediated interaction energy.

RestrictedBoltzmannMachines.log_likelihood — Function

log_likelihood(rbm, v, β=1)

Log-likelihood of v under rbm, with the partition function compued by extensive enumeration. For discrete layers, this is exponentially slow for large machines.

RestrictedBoltzmannMachines.log_partition — Function

log_partition(rbm, β = 1)

Log-partition of the rbm at inverse temperature β, computed by extensive enumeration of visible states (except for particular cases such as Gaussian-Gaussian RBM). This is exponentially slow for large machines.

If your RBM has a smaller hidden layer, consider using flip_layers.

RestrictedBoltzmannMachines.log_pseudolikelihood — Function

log_pseudolikelihood(rbm, v, β=1; exact=false)

Log-pseudolikelihood of v. If exact is true, the exact pseudolikelihood is returned. But this is slow if v consists of many samples. Therefore by default exact is false, in which case the result is a stochastic approximation, where a random site is selected for each sample, and its conditional probability is calculated. In average the results with exact = false coincide with the deterministic result, and the estimate is more precise as the number of samples increases.

RestrictedBoltzmannMachines.log_pseudolikelihood_exact — Function

log_pseudolikelihood_exact(rbm, v, β = 1)

Log-pseudolikelihood of v. This function computes the exact pseudolikelihood, doing traces over all sites. Note that this can be slow for large number of samples.

RestrictedBoltzmannMachines.log_pseudolikelihood_sites — Function

log_pseudolikelihood_sites(rbm, v, sites, β=1)

Log-pseudolikelihood of a site conditioned on the other sites, where sites is an array of site indices (CartesianIndex), one for each sample. Returns an array of log-pseudolikelihood values, for each sample.

RestrictedBoltzmannMachines.log_pseudolikelihood_stoch — Function

log_pseudolikelihood_stoch(rbm, v, β=1)

Log-pseudolikelihood of v. This function computes an stochastic approximation, by doing a trace over random sites for each sample. For large number of samples, this is in average close to the exact value of the pseudolikelihood.

RestrictedBoltzmannMachines.lognormcdf — Method

lognormcdf(a, b)

Computes log(normcdf(a, b)), but retaining accuracy.

RestrictedBoltzmannMachines.mean_ — Method

mean_(A; dims)

Takes the mean of A across dimensions dims and drops them.

RestrictedBoltzmannMachines.mills — Method

mills(x::Real)

Mills ratio of the standard normal distribution. Defined as (1 - cdf(x)) / pdf(x).

RestrictedBoltzmannMachines.minibatch_count — Method

minibatch_count(nobs; batchsize)

Number of minibatches.

RestrictedBoltzmannMachines.minibatch_count — Method

minibatch_count(data; batchsize)

Number of minibatches.

RestrictedBoltzmannMachines.minibatches — Method

minibatches(nobs; batchsize, shuffle = true)

Partition nobs into minibatches of length n. If necessary repeats some observations to complete last batches. (Therefore all batches are of the same size n).

RestrictedBoltzmannMachines.minibatches — Method

minibatches(datas...; batchsize)

Splits the given datas into minibatches. Each minibatch is a tuple where each entry is a minibatch from the corresponding data within datas. All minibatches are of the same size batchsize (if necessary repeating some samples at the last minibatches).

RestrictedBoltzmannMachines.normalize_weights — Method

normalize_weights(rbm)

Rescales weights so that norm(w[:,μ]) = 1 for all μ (making individual weights ~ 1/√N). The resulting RBM shares layer parameters with the original, but weights are a new array.

RestrictedBoltzmannMachines.normcdf — Method

normcdf(a, b)

Probablity that a ≤ Z ≤ b, where Z is a standard normal samplefrominputs variable. WARNING: Silently returns a negative value if a > b.

RestrictedBoltzmannMachines.normcdf — Method

normcdf(x)

Probablity that Z ≤ x, where Z is a standard normal samplefrominputs variable.

RestrictedBoltzmannMachines.normcdfinv — Method

normcdfinv(x)

Inverse of normcdf.

RestrictedBoltzmannMachines.onehot_decode — Method

onehot_decode(X)

Given a onehot encoded array X of N + 1 dimensions, returns the equivalent categorical array of N dimensions.

RestrictedBoltzmannMachines.onehot_encode — Function

onehot_encode(A, code)

Given an array A of N dimensions, returns a one-hot encoded BitArray of N + 1 dimensions where single entries of the first dimension are one.

RestrictedBoltzmannMachines.randgumbel — Method

randgumbel(T = Float64)

Generates a random Gumbel variate.

RestrictedBoltzmannMachines.randnt — Method

randnt([rng], a)

Random standard normal lower truncated at a (that is, Z ≥ a).

RestrictedBoltzmannMachines.randnt_half — Method

randnt_half([rng], μ, σ)

Samples the normal distribution with mean μ and standard deviation σ truncated to positive values.

RestrictedBoltzmannMachines.reconstruction_error — Function

reconstruction_error(rbm, v, β = 1; steps = 1)

Stochastic reconstruction error of v.

RestrictedBoltzmannMachines.sample_from_inputs — Function

sample_from_inputs(layer, inputs, β = 1)

Samples layer configurations conditioned on inputs.

RestrictedBoltzmannMachines.sample_h_from_h — Function

sample_h_from_h(rbm, h, β = 1; steps = 1)

Samples a hidden configuration conditional on another hidden configuration h.

RestrictedBoltzmannMachines.sample_h_from_v — Function

sample_h_from_v(rbm, v, β=1)

Samples a hidden configuration conditional on the visible configuration v.

RestrictedBoltzmannMachines.sample_v_from_h — Function

sample_v_from_h(rbm, h, β = 1)

Samples a visible configuration conditional on the hidden configuration h.

RestrictedBoltzmannMachines.sample_v_from_v — Function

sample_v_from_v(rbm, v, β = 1; steps = 1)

Samples a visible configuration conditional on another visible configuration v.

RestrictedBoltzmannMachines.sqrt1half — Method

sqrt1half(x)

Accurate computation of sqrt(1 + (x/2)^2) + |x|/2.

RestrictedBoltzmannMachines.std_ — Method

std_(A; dims)

Takes the standard deviation of A across dimensions dims and drops them.

RestrictedBoltzmannMachines.substitution_matrix_exhaustive — Function

substitution_matrix_exhaustive(rbm, v, β = 1)

Returns an q x N x B tensor of free energies F, where q is the number of possible values of each site, B the number of data points, and N the sequence length:

`q, N, B = size(v)

Thus F and v have the same size. The entry F[x,i,b] gives the free energy cost of flipping site i to x of v[b] from its original value to x, that is:

F[x,i,b] = free_energy(rbm, v_, β) - free_energy(rbm, v[b], β)

where v_ is the same as v[b] in all sites but i, where v_ has the value x.

Note that i can be a set of indices.

RestrictedBoltzmannMachines.substitution_matrix_sites — Function

substitution_matrix_sites(rbm, v, sites, β = 1)

Returns an q x B matrix of free energies F, where q is the number of possible values of each site, and B the number of data points. The entry F[x,b] equals the free energy cost of flipping site[b] of v[b] to x, that is (schemetically):

F[x, b] = free_energy(rbm, v_) - free_energy(rbm, v)

where v = v[b], and v_ is the same as v in all sites except site[b], where v_ has the value x.

RestrictedBoltzmannMachines.sum_ — Method

sum_(A; dims)

Sums A over dimensions dims and drops them.

RestrictedBoltzmannMachines.tnmean — Method

tnmean(a)

Mean of the standard normal distribution, truncated to the interval (a, +∞).

RestrictedBoltzmannMachines.tnstd — Method

tnstd(a)

Standard deviation of the standard normal distribution, truncated to the interval (a, +∞).

RestrictedBoltzmannMachines.tnvar — Method

tnvar(a)

Variance of the standard normal distribution, truncated to the interval (a, +∞). WARNING: Fails for very very large values of a.

RestrictedBoltzmannMachines.train! — Method

train!(rbm, data)

Trains the RBM on data.

RestrictedBoltzmannMachines.tuplen — Method

tuplen(Val(N))

Constructs the tuple (1, 2, ..., N).

RestrictedBoltzmannMachines.unwhiten — Method

unwhiten(rbm, data)

Given an RBM trained on whitened data, returns an RBM that can look at original data.

RestrictedBoltzmannMachines.var_ — Method

var_(A; dims)

Takes the variance of A across dimensions dims and drops them.

RestrictedBoltzmannMachines.weighted_mean — Method

weighted_mean(v, w)

Mean of v with weights w.

RestrictedBoltzmannMachines.zerosum! — Method

zerosum!(rbm)

If the rbm has Potts layers (visible or hidden), fixes zerosum gauge on the weights and on the layer fields. Otherwise, does nothing.