Reference

RestrictedBoltzmannMachines.Binary — Type

Binary(θ)

Binary layer, with external fields θ.

RestrictedBoltzmannMachines.Gaussian — Type

Gaussian(θ, γ)

Gaussian layer, with location parameters θ and scale parameters γ.

RestrictedBoltzmannMachines.Potts — Type

Potts(θ)

Potts layer, with external fields θ. Encodes categorical variables as one-hot vectors. The number of classes is the size of the first dimension.

RestrictedBoltzmannMachines.RBM — Type

RBM{V,H,W}

RBM, with visible layer of type V, hidden layer of type H, and weights of type W.

RestrictedBoltzmannMachines.ReLU — Type

ReLU(θ, γ)

ReLU layer, with location parameters θ and scale parameters γ.

RestrictedBoltzmannMachines.Spin — Type

Spin(θ)

Spin layer, with external fields θ. The energy of a layer with units $s_i$ is given by:

\[E = -\sum_i \theta_i s_i\]

where each spin $s_i$ takes values $\pm 1$.

RestrictedBoltzmannMachines.BinaryRBM — Method

BinaryRBM(a, b, w)
BinaryRBM(N, M)

Construct an RBM with binary visible and hidden units, which has an energy function:

\[E(v, h) = -a'v - b'h - v'wh\]

Equivalent to RBM(Binary(a), Binary(b), w).

RestrictedBoltzmannMachines.HopfieldRBM — Method

HopfieldRBM(g, θ, γ, w)
HopfieldRBM(g, w)

Construct an RBM with spin visible units and Gaussian hidden units. If not given, θ = 0 and γ = 1 by default.

\[E(v, h) = -g'v - θ'h + \sum_\mu \frac{γ_\mu}{2} h_\mu^2 - v'wh\]

RestrictedBoltzmannMachines.ais — Method

ais(rbm; nbetas = 10000, nsamples = 10, init = visible(rbm))

Estimates the log-partition function of rbm by Annealed Importance Sampling (AIS). Here init is an initial independent-site distribution, represented by a visible layer. Returns nsamples variates, which can be averaged.

If R = ais(rbm), then mean(exp.(R)) is an unbiased estimator of the partition function. One can use mean(R) or logmeanexp(R) to estimate the log-partition function, and these estimators have different properties. std(R) or logstdexp(R) can be used to get (rough) estimates of the sampling error, but do not trust these because the estimator is biased for any number of finite interpolating temperatures!

RestrictedBoltzmannMachines.anneal — Method

anneal(visible_init, rbm_final; β)

Returns an RBM that interpolates between an independent-site model visible_init and rbm. Denoting by E0(v) the energies assigned by visible_init, and by E1(v, h) the energies assigned by rbm_final, the returned RBM assigns energies given by:

E(v,h) = (1 - β) * E0(v) + β * E1(v, h)

RestrictedBoltzmannMachines.batch_size — Method

batch_size(layer, x)

Batch sizes of x, with respect to layer.

RestrictedBoltzmannMachines.batch_size — Method

batch_size(rbm, v, h)

Returns the batch size if energy(rbm, v, h) were computed.

RestrictedBoltzmannMachines.batchcov — Method

batchcov(layer, x; wts = nothing, [mean])

Covariance of x over batch dimensions, weigthed by wts.

RestrictedBoltzmannMachines.batchdims — Method

batchdims(layer, x)

Indices of batch dimensions in x, with respect to layer.

RestrictedBoltzmannMachines.batchmean — Method

batchmean(layer, x; wts = nothing)

Mean of x over batch dimensions, weigthed by wts.

RestrictedBoltzmannMachines.batchstd — Method

batchstd(layer, x; wts = nothing, [mean])

Standard deviation of x over batch dimensions, weigthed by wts.

RestrictedBoltzmannMachines.batchvar — Method

batchvar(layer, x; wts = nothing, [mean])

Variance of x over batch dimensions, weigthed by wts.

RestrictedBoltzmannMachines.block_matrix_invert — Method

block_matrix_invert(A, B, C, D)

Inversion of a block matrix, using the formula:

\[\begin{bmatrix} \mathbf{A} & \mathbf{B} \\ \mathbf{C} & \mathbf{D} \end{bmatrix}^{-1} = \begin{bmatrix} \left(\mathbf{A} - \mathbf{B} \mathbf{D}^{-1} \mathbf{C}\right)^{-1} & \mathbf{0} \\ \mathbf{0} & \left(\mathbf{D} - \mathbf{C} \mathbf{A}^{-1} \mathbf{B}\right)^{-1} \end{bmatrix} \begin{bmatrix} \mathbf{I} & -\mathbf{B} \mathbf{D}^{-1} \\ -\mathbf{C} \mathbf{A}^{-1} & \mathbf{I} \end{bmatrix}\]

Assumes that A and D are square and invertible.

RestrictedBoltzmannMachines.block_matrix_logdet — Method

block_matrix_logdet(A, B, C, D)

Log-determinant of a block matrix using the determinant lemma.

\[\det\left( \begin{bmatrix} \mathbf{A} & \mathbf{B} \\ \mathbf{C} & \mathbf{D} \end{bmatrix} \right) = \det(A) \det(D - CA^{-1}B) = \det(D) \det(A - BD^{-1}C)\]

Here we assume that A and D are invertible, and moreover are easy to invert (for example, if they are diagonal). We use this to chose one or the other of the two formulas above.

RestrictedBoltzmannMachines.broadlike — Method

broadlike(A, B...)

Broadcasts A into the size of A .+ B .+ ... (without actually doing a sum).

RestrictedBoltzmannMachines.categorical_rand — Method

categorical_rand(ps)

Randomly draw i with probability ps[i]. You must ensure that ps defines a proper probability distribution.

RestrictedBoltzmannMachines.categorical_sample — Method

categorical_sample(P)

Given a probability array P of size (q, *), returns an array C of size (*), such that C[i] ∈ 1:q is a random sample from the categorical distribution P[:,i]. You must ensure that P defines a proper probability distribution.

RestrictedBoltzmannMachines.categorical_sample_from_logits — Method

categorical_sample_from_logits(logits)

Given a logits array logits of size (q, *) (where q is the number of classes), returns an array X of size (*), such that X[i] is a categorical random sample from the distribution with logits logits[:,i].

RestrictedBoltzmannMachines.categorical_sample_from_logits_gumbel — Method

categorical_sample_from_logits_gumbel(logits)

Like categoricalsamplefrom_logits, but using the Gumbel trick.

RestrictedBoltzmannMachines.cd! — Method

cd!(rbm, data)

Trains the RBM on data using contrastive divergence.

RestrictedBoltzmannMachines.cdad! — Method

cdad!(rbm, data)

Trains the RBM on data using contrastive divergence. Computes gradients with Zygote.

RestrictedBoltzmannMachines.center_gradient — Method

center_gradient(rbm, ∂, λv, λh)

Given a gradient ∂ of rbm, returns the gradient of the equivalent centered RBM with offset_v and offset_h.

RestrictedBoltzmannMachines.cold_metropolis — Method

cold_metropolis(rbm, v; steps = 1)

Samples the rbm at zero temperature, starting from configuration v.

RestrictedBoltzmannMachines.collect_states — Method

collect_states(layer)

Returns an array of all states of layer. Only defined for discrete layers.

Warning

Use only for small layers. For large layers, the exponential number of states will not fit in memory.

RestrictedBoltzmannMachines.colors — Method

colors(layer)

Number of possible states of units in discrete layers.

RestrictedBoltzmannMachines.contrastive_divergence — Method

contrastive_divergence(rbm, vd, vm; wd = 1, wm = 1)

Contrastive divergence loss. vd is a data sample, and vm are samples from the model.

RestrictedBoltzmannMachines.default_optimizer — Method

default_optimizer(nsamples, batchsize, epochs; decay_after, decay_final, clip, optim)

Sane defaults for optimizer. All keyword arguments are optional and have default settings. Trains with constant learning rate for first decay_after period, then decays learning rate exponentially every epoch, starting after decay_after of training time, until reaching decay_final at the last epoch. Clips gradients by clip.

RestrictedBoltzmannMachines.energies — Method

energies(layer, x)

Energies of units in layer (not reduced over layer dimensions).

RestrictedBoltzmannMachines.energy — Method

energy(layer, x)

Layer energy, reduced over layer dimensions.

RestrictedBoltzmannMachines.energy — Method

energy(rbm, v, h)

Energy of the rbm in the configuration (v,h).

RestrictedBoltzmannMachines.flatten — Method

flatten(layer, x)

Returns a vectorized version of x.

RestrictedBoltzmannMachines.fpcd! — Method

fpcd!(rbm, data)

Trains the RBM on data using Persistent Contrastive divergence, with fast weights. See http://dl.acm.org/citation.cfm?id=1553374.1553506.

RestrictedBoltzmannMachines.free_energy — Function

free_energy(layer, inputs = 0)

Cumulant generating function of layer, reduced over layer dimensions.

RestrictedBoltzmannMachines.free_energy — Method

free_energy(rbm, v)

Free energy of visible configuration (after marginalizing hidden configurations).

RestrictedBoltzmannMachines.generate_sequences — Function

generate_sequences(n, A = 0:1)

Retruns an iterator over all sequences of length n out of the alphabet A.

RestrictedBoltzmannMachines.gradmult — Method

gradmult(∂, λ)

Multiplies gradients by a scalar λ.

RestrictedBoltzmannMachines.gradnorms — Method

gradnorms(∂)

Computes gradient norms.

RestrictedBoltzmannMachines.initialize! — Function

initialize!(rbm, [data]; ϵ = 1e-6)

Initializes the RBM and returns it. If provided, matches average visible unit activities from data.

initialize!(layer, [data]; ϵ = 1e-6)

Initializes a layer and returns it. If provided, matches average unit activities from data.

RestrictedBoltzmannMachines.initialize_w! — Method

initialize_w!(rbm, data; λ = 0.1)

Initializes rbm.w such that typical inputs to hidden units are λ.

RestrictedBoltzmannMachines.inputs_h_to_v — Method

inputs_h_to_v(rbm, h)

Interaction inputs from hidden to visible layer.

RestrictedBoltzmannMachines.inputs_v_to_h — Method

inputs_v_to_h(rbm, v)

Interaction inputs from visible to hidden layer.

RestrictedBoltzmannMachines.interaction_energy — Method

interaction_energy(rbm, v, h)

Weight mediated interaction energy.

RestrictedBoltzmannMachines.log_likelihood — Method

log_likelihood(rbm, v)

Log-likelihood of v under rbm, with the partition function compued by extensive enumeration. For discrete layers, this is exponentially slow for large machines.

RestrictedBoltzmannMachines.log_partition — Method

log_partition(rbm)

Log-partition of rbm, computed by extensive enumeration of visible states (except for particular cases such as Gaussian-Gaussian RBM). This is exponentially slow for large machines.

If your RBM has a smaller hidden layer, mirroring the layers of the rbm first (see mirror).

RestrictedBoltzmannMachines.log_partition_zero_weight — Method

log_partition_zero_weight(rbm)

Log-partition function of a zero-weight version of rbm.

RestrictedBoltzmannMachines.log_pseudolikelihood — Method

log_pseudolikelihood(rbm, v; exact = false)

Log-pseudolikelihood of v. If exact is true, the exact pseudolikelihood is returned. But this is slow if v consists of many samples. Therefore by default exact is false, in which case the result is a stochastic approximation, where a random site is selected for each sample, and its conditional probability is calculated. In average the results with exact = false coincide with the deterministic result, and the estimate is more precise as the number of samples increases.

RestrictedBoltzmannMachines.log_pseudolikelihood_exact — Method

log_pseudolikelihood_exact(rbm, v)

Log-pseudolikelihood of v. This function computes the exact pseudolikelihood, doing traces over all sites. Note that this can be slow for large number of samples.

RestrictedBoltzmannMachines.log_pseudolikelihood_sites — Method

log_pseudolikelihood_sites(rbm, v, sites)

Log-pseudolikelihood of a site conditioned on the other sites, where sites is an array of site indices (CartesianIndex), one for each sample. Returns an array of log-pseudolikelihood values, for each sample.

RestrictedBoltzmannMachines.log_pseudolikelihood_stoch — Method

log_pseudolikelihood_stoch(rbm, v)

Log-pseudolikelihood of v. This function computes an stochastic approximation, by doing a trace over random sites for each sample. For large number of samples, this is in average close to the exact value of the pseudolikelihood.

RestrictedBoltzmannMachines.logmeanexp — Method

logmeanexp(A; dims=:)

Computes log.(mean(exp.(A); dims)), in a numerically stable way.

RestrictedBoltzmannMachines.logstdexp — Method

logstdexp(A; dims=:)

Computes log.(std(exp.(A); dims)), in a numerically stable way.

RestrictedBoltzmannMachines.logvarexp — Method

logvarexp(A; dims=:)

Computes log.(var(exp.(A); dims)), in a numerically stable way.

RestrictedBoltzmannMachines.mean_h_from_v — Method

mean_h_from_v(rbm, v)

Mean unit activation values, conditioned on the other layer, <h | v>.

RestrictedBoltzmannMachines.mean_v_from_h — Method

mean_v_from_h(rbm, v)

Mean unit activation values, conditioned on the other layer, <v | h>.

RestrictedBoltzmannMachines.metropolis! — Method

metropolis!(v, rbm; β = 1)

Metropolis-Hastings sampling from rbm at inverse temperature β. Uses v[:,:,..,:,1] as initial configurations, and writes the Monte-Carlo chains in v[:,:,..,:,2:end].

RestrictedBoltzmannMachines.metropolis — Method

metropolis(rbm, v; β = 1, steps = 1)

Metropolis-Hastings sampling from rbm at inverse temperature β, starting from configuration v.

RestrictedBoltzmannMachines.minibatch_count — Method

minibatch_count(nobs; batchsize)

Number of minibatches.

RestrictedBoltzmannMachines.minibatch_count — Method

minibatch_count(data; batchsize)

Number of minibatches.

RestrictedBoltzmannMachines.minibatches — Method

minibatches(nobs; batchsize, shuffle = true)

Partition nobs into minibatches of length n. If necessary repeats some observations to complete last batches. (Therefore all batches are of the same size n).

RestrictedBoltzmannMachines.minibatches — Method

minibatches(datas...; batchsize)

Splits the given datas into minibatches. Each minibatch is a tuple where each entry is a minibatch from the corresponding data within datas. All minibatches are of the same size batchsize (if necessary repeating some samples at the last minibatches).

RestrictedBoltzmannMachines.mirror — Method

mirror(rbm)

Returns a new RBM with viible and hidden layers flipped.

RestrictedBoltzmannMachines.mode_h_from_v — Method

mode_h_from_v(rbm, v)

Mode unit activations, conditioned on the other layer.

RestrictedBoltzmannMachines.mode_v_from_h — Method

mode_v_from_h(rbm, h)

Mode unit activations, conditioned on the other layer.

RestrictedBoltzmannMachines.onehot_decode — Method

onehot_decode(X)

Given a onehot encoded array X of N + 1 dimensions, returns the equivalent categorical array of N dimensions.

RestrictedBoltzmannMachines.onehot_encode — Function

onehot_encode(A, code)

Given an array A of N dimensions, returns a one-hot encoded BitArray of N + 1 dimensions where single entries of the first dimension are one.

RestrictedBoltzmannMachines.pcd! — Method

pcd!(rbm, data)

Trains the RBM on data using Persistent Contrastive divergence.

RestrictedBoltzmannMachines.pcd_bnorm! — Method

pcd_bnorm!(rbm, data)

Trains the RBM on data using Persistent Contrastive divergence, with centered gradients, as done in the PGM repo, https://github.com/jertubiana/PGM.

This is almost equivalent to centering visible units only (see https://github.com/cossio/CenteredRBMs.jl), but the centering of the hidden unit parameters is done much smoother.

RestrictedBoltzmannMachines.rais — Method

rais(rbm, v; nbetas = 10000, nsamples = 10, init = visible(rbm))

Estimates the log-partition function of rbm by reverse Annealed Importance Sampling (AIS). These requires a set of samples v from the rbm. In pratice, we can use data points, assuming the rbm has been well-trained and approximates well the empirical distribution. As in ais, init is an initial independent-site distribution, represented by a visible layer. It is recommended to select init to match the independent-site statistics of the data (so try to not use the default value visible(rbm)!). Returns nsamples variates, which can be averaged.

If R = rais(rbm), then mean(exp.(-R)) is an unbiased estimator of 1/Z, the inverse of the partition function. In practice, one can use mean(R) or logmeanexp(R) to estimate the log-partition function, but note that these estimators have different properties. std(R) or logstdexp(R) can be used to get (rough) estimates of the sampling error, but do not trust these because the estimator is biased for any number of finite interpolating temperatures!

RestrictedBoltzmannMachines.randgumbel — Method

randgumbel(T = Float64)

Generates a random Gumbel variate.

RestrictedBoltzmannMachines.randnt — Method

randnt([rng], a)

Random standard normal lower truncated at a (that is, Z ≥ a).

RestrictedBoltzmannMachines.randnt_half — Method

randnt_half([rng], μ, σ)

Samples the normal distribution with mean μ and standard deviation σ truncated to positive values.

RestrictedBoltzmannMachines.rdm! — Method

rdm!(rbm, data)

Trains the RBM on data using contrastive divergence with randomly initialized chains. See http://arxiv.org/abs/2105.13889.

RestrictedBoltzmannMachines.reconstruction_error — Method

reconstruction_error(rbm, v; steps = 1)

Stochastic reconstruction error of v.

RestrictedBoltzmannMachines.repeat_size — Method

repeat_size(sz, r...)

Returns size(repeat(A, r...)), provided size(A) == sz.

RestrictedBoltzmannMachines.rescale_activations! — Method

rescale_activations!(layer, λ::AbstractArray)

For continuous layers with scale parameters, re-parameterizes such that unit activations are divided by λ, and returns true. For other layers just returns false.

RestrictedBoltzmannMachines.rescale_hidden! — Method

rescale_hidden!(rbm, λ::AbstractArray)

For continuous hidden units with a scale parameter, scales parameters such that hidden unit activations are divided by λ. For other hidden units does nothing. The resulting RBM is equivalent to the original one.

RestrictedBoltzmannMachines.reshape_maybe — Method

reshape_maybe(x, shape)

Like reshape(x, shape), except that zero-dimensional outputs are returned as scalars.

RestrictedBoltzmannMachines.sample_h_from_h — Method

sample_h_from_h(rbm, h; steps = 1)

Samples a hidden configuration conditional on another hidden configuration h. Ensures type stability by requiring that the returned array is of the same type as h.

RestrictedBoltzmannMachines.sample_h_from_v — Method

sample_h_from_v(rbm, v)

Samples a hidden configuration conditional on the visible configuration v.

RestrictedBoltzmannMachines.sample_v_from_h — Method

sample_v_from_h(rbm, h)

Samples a visible configuration conditional on the hidden configuration h.

RestrictedBoltzmannMachines.sample_v_from_v — Method

sample_v_from_v(rbm, v; steps = 1)

Samples a visible configuration conditional on another visible configuration v. Ensures type stability by requiring that the returned array is of the same type as v.

RestrictedBoltzmannMachines.sitedims — Method

sitedims(layer)

Number of dimensions of layer, with special handling of Potts layer, for which the first dimension doesn't count as a site dimension.

RestrictedBoltzmannMachines.sitesize — Method

sitesize(layer)

Size of layer, with special handling of Potts layer, for which the first dimension doesn't count as a site dimension.

RestrictedBoltzmannMachines.sqrt1half — Method

sqrt1half(x)

Accurate computation of sqrt(1 + (x/2)^2) + |x|/2.

RestrictedBoltzmannMachines.substitution_matrix_exhaustive — Function

substitution_matrix_exhaustive(rbm, v)

Returns an q x N x B tensor of free energies F, where q is the number of possible values of each site, B the number of data points, and N the sequence length:

`q, N, B = size(v)

Thus F and v have the same size. The entry F[x,i,b] gives the free energy cost of flipping site i to x of v[b] from its original value to x, that is:

F[x,i,b] = free_energy(rbm, v_) - free_energy(rbm, v[b])

where v_ is the same as v[b] in all sites but i, where v_ has the value x.

Note that i can be a set of indices.

RestrictedBoltzmannMachines.substitution_matrix_sites — Function

substitution_matrix_sites(rbm, v, sites)

Returns an q x B matrix of free energies F, where q is the number of possible values of each site, and B the number of data points. The entry F[x,b] equals the free energy cost of flipping site[b] of v[b] to x, that is (schemetically):

F[x, b] = free_energy(rbm, v_) - free_energy(rbm, v)

where v = v[b], and v_ is the same as v in all sites except site[b], where v_ has the value x.

RestrictedBoltzmannMachines.suffstats — Function

suffstats(layer, data; [wts])

Computes the sufficient statistics of layer extracted from data.

RestrictedBoltzmannMachines.tnmean — Method

tnmean(a)

Mean of the standard normal distribution, truncated to the interval (a, +∞).

RestrictedBoltzmannMachines.tnmeanvar — Method

tnmeanvar(a)

Mean and variance of the standard normal distribution truncated to the interval (a, +∞). Equivalent to tnmean(a), tnvar(a) but saves some common computations. WARNING: tnvar(a) can fail for very very large values ofa`.

RestrictedBoltzmannMachines.tnvar — Method

tnvar(a)

Variance of the standard normal distribution, truncated to the interval (a, +∞). WARNING: Fails for very very large values of a.

RestrictedBoltzmannMachines.uncenter_step — Method

uncenter_step(rbm, ∂, λv, λh)

Given parameter update step ∂ of a centered rbm with offset_v and offset_h, returns the corresponding gradient of the equivalent uncentered RBM.

RestrictedBoltzmannMachines.update! — Method

update!(∂, x, optim)

Computes parameter update step according to optim algorithm (e.g. ADAM), and gradient ∂ of parameters x. Overwrites ∂ with update step and returns it. Note that this does not update parameters.

RestrictedBoltzmannMachines.update! — Method

update!(x, ∂)

Updates parameters according to steps ∂.

RestrictedBoltzmannMachines.var_h_from_v — Method

var_h_from_v(rbm, v)

Variance of unit activation values, conditioned on the other layer, var(h | v).

RestrictedBoltzmannMachines.var_v_from_h — Method

var_v_from_h(rbm, v)

Variance of unit activation values, conditioned on the other layer, var(v | h).

RestrictedBoltzmannMachines.wmean — Method

wmean(A; wts = nothing, dims = :)

Weighted mean of A along dimensions dims, weighted by wts.

\[\frac{\sum_i A_i w_i}{\sum_i w_i}\]

RestrictedBoltzmannMachines.wsum — Method

wsum(A; wts = nothing, dims = :)

Weighted sum of A along dimensions dims, weighted by wts.

\[\frac{\sum_i A_i w_i}\]

RestrictedBoltzmannMachines.zerosum! — Method

zerosum!(∂, rbm)

Enforces zerosum gauge on the RBM gradient.

RestrictedBoltzmannMachines.zerosum! — Method

zerosum!(rbm)

Fixes zerosum gauge for Potts layers (visible or hidden). If the rbm doesn't have Potts layers, does nothing.

RestrictedBoltzmannMachines.zerosum — Method

zerosum(rbm)

Fixes zerosum gauge for Potts layers (visible or hidden). If the rbm doesn't have Potts layers, does nothing.

RestrictedBoltzmannMachines.∂energy — Method

∂energy(layer, data; wts = nothing)
∂energy(layer, stats)

Derivative of average energy of data with respect to layer parameters. In the second form, stats are the pre-computed sufficient statistics of the layer. See suffstats.

RestrictedBoltzmannMachines.∂free_energy — Function

∂free_energy(layer, inputs = 0; wts = 1)

Unit activation moments, conjugate to layer parameters. These are obtained by differentiating free_energies with respect to the layer parameters. Averages over configurations (weigthed by wts).

RestrictedBoltzmannMachines.∂free_energy — Method

∂free_energy(rbm, v)

Gradient of free_energy(rbm, v) with respect to model parameters. If v consists of multiple samples (batches), then an average is taken.

RestrictedBoltzmannMachines.∂regularize! — Method

∂regularize!(∂, rbm; l2_fields = 0, l1_weights = 0, l2_weights = 0, l2l1_weights = 0)

Updates RBM gradients ∂, with the regularization gradient.